api/requirements.txt
django==2.0
djangorestframework
django-filter
psycopg2
uwsgi
